+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
   manner ([#154](https://gitlab.heigit.org/climate-action/climatoology/-/issues/154))
 - a teaser sentence for plugins can now be specified to be displayed on hover in the
   front-end ([#162](https://gitlab.heigit.org/climate-action/climatoology/-/issues/162)).
+- a user agent (application name) to the backand database connection
 
 ## [6.3.1](https://gitlab.heigit.org/climate-action/climatoology/-/releases/6.3.1) - 2025-04-28